fix type conversion for max wait
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@13969 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/java/org/gusdb/wdk/model/config/ModelConfigDB.java
@@ -176,7 +176,7 @@ public abstract class ModelConfigDB implements ConnectionPoolConfig {
    * @param maxWait
    *          the maxWait to set
    */
-  public void setMaxWait(short maxWait) {
+  public void setMaxWait(long maxWait) {
     this.maxWait = maxWait;
   }
